Run tests always with vim-test, and use neoterm strategy in neovim
Instead of using `vim-spec-runner` and `vim-rails` test runners in different
scenarios, now we'll use always `vim-test`, so `vim-spec-runner` can be
removed (nont `vim-rails` which provides much more).

When neovim is not installed we'll use `dispatch` strategy, as before.
But when `neovim` is installed we'll use `neoterm` strategy, provided by
`neoterm` plugin, which improves `neovim` terminal and rides off some odd
behaviour when used as a `vim-test` strategy.

This is just an experiment. Let's see if it is comfortable or we revert
it...